+++ app/mainwindow.cpp
@@ -8,7 +8,7 @@
 #include <QStyle>
 #include <QEvent>
 
-MainWindow::MainWindow(const QStringList &args, Config::Local &local_c, Config::Global &global_c, QWidget *parent) :
+MainWindow::MainWindow(const QStringList &args, IPC::Instance *instance, Config::Local &local_c, Config::Global &global_c, QWidget *parent) :
   QMainWindow(parent), ui(new Ui::MainWindow), local_conf(local_c), global_conf(global_c) {
   trayicon = nullptr;
 #if defined(MPRIS_ENABLE)
@@ -72,6 +72,13 @@ MainWindow::MainWindow(const QStringList &args, Config::Local &local_c, Config::
   setupSortMenu();
 
   preloadPlaylist(args);
+
+  connect(instance, &IPC::Instance::load_files_received, [=](const QStringList &lst) {
+    preloadPlaylist(lst);
+    show();
+    raise();
+    setFocus();
+  });
 }
 
 MainWindow::~MainWindow() {
@@ -266,6 +273,8 @@ void MainWindow::setupPlaybackDispatch() {
   connect(player, &Playback::Controller::nextRequested, dispatch, &Playback::Dispatch::on_nextRequested);
   connect(player, &Playback::Controller::startRequested, dispatch, &Playback::Dispatch::on_startRequested);
   connect(dispatch, &Playback::Dispatch::play, player, &Playback::Controller::play);
+
+  connect(playlists, &PlaylistsUi::Controller::doubleclicked, dispatch, &Playback::Dispatch::on_startFromPlaylistRequested);
 }
 
 void MainWindow::setupStatusBar() {